# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчёта трудовых норм и учёта производственных заказов",
        "title": "Сервис нормирования API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/admin//coefficient/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Обновить соэффициент",
                "parameters": [
                    {
                        "description": "Список коэффициентов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.CoefficientPEOAdmin"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "ошибка парсинга JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/coefficient": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить коэффициенты",
                "responses": {
                    "200": {
                        "description": "Список коэффициентов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.CoefficientPEOAdmin"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/employees": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить сотрудников",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.EmployeesAdmin"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/employees/teams": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить бригады",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.TeamAdmin"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/employees/update/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Обновить сотрудника",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сотрудника",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные сотрудника",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UpdateEmployeeInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Неверный JSON или ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/template/new": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Сохранить новый шаблон",
                "parameters": [
                    {
                        "description": "Данные нового шаблона",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/save.TemplateAdminSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ошибка парсинга JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/report/excel": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "report-excel"
                ],
                "summary": "Скачать отчёт в формате Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала периода (YYYY-MM-DD), по умолчанию — начало текущего месяца",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода (YYYY-MM-DD), по умолчанию — текущая дата",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Номер заказа",
                        "name": "order_num",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Типы изделий (можно указать несколько)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel-файл с отчётом",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "invalid from date / invalid to date",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "save.TemplateAdminSaveRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "head_name": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Operation"
                    }
                },
                "profile": {
                    "type": "string"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Rule"
                    }
                },
                "systema": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                }
            }
        },
...
        "storage.CoefficientPEOAdmin": {
            "type": "object",
            "properties": {
                "coefficient": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "storage.EmployeesAdmin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.TeamAdmin"
                    }
                }
            }
        },
...
        "storage.Operation": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "depends_on": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "formula": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "minutes": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "show_calc": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "storage.Rule": {
            "type": "object",
            "properties": {
                "condition": {
                    "type": "object",
                    "additionalProperties": true
                },
                "minutes": {
                    "description": "fallback",
                    "type": "number"
                },
                "minutes_per_unit": {
                    "type": "number"
                },
                "mode": {
                    "description": "\"set\", \"multiplied\", \"additive\"",
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "unitField": {
                    "type": "string"
                },
                "value": {
                    "description": "fallback",
                    "type": "number"
                },
                "value_per_unit": {
                    "type": "number"
                }
            }
        },
...
        "storage.TeamAdmin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "storage.UpdateEmployeeInput": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "team_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}